pyproject.toml
[project]
requires-python = ">=3.11"

[tool.poetry]
name = "enka-card-api"
version = "1.3.1"
description = "API to generate Enka cards."
authors = ["seriaati <seria.ati@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pillow = "^10.2.0"
cachetools = "^5.3.3"
asyncache = "^0.3.1"
numpy = "^1.26.4"
more-itertools = "^10.2.0"
aiofiles = "^23.2.1"
pydantic = "v1.10.9"
matplotlib = "^3.8.3"
python-magic = "^0.4.27"
beautifulsoup4 = "^4.12.3"
uvicorn = "^0.27.1"
requests = "^2.31.0"
enkacard = "^3.3.7"
enkanetworkv2-py = {git = "https://github.com/seriaati/EnkaNetworkV2.py"}
starrailcard = "^2.1.8"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
pre-commit = "^3.6.1"

[tool.poetry.group.linux]
optional = true

[tool.poetry.group.linux.dependencies]
uvloop = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E","W","C90","F","UP","B","SIM","I","N","TCH",
    "ANN","ASYNC","A","C4","EM","FA","ICN","G",
    "PIE","T20","ARG","ERA","LOG","PL","TRY"
]
ignore = [
    "E501","ANN401","ANN101","ANN102","ANN003",
    "PLR0913","PLR2004","PLR0917","PLR6301",
    "ANN002"
]
preview = true

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F403"]
"playground.*.py" = ["ALL"]
"EnkaCard/**/*.py" = ["ALL"]
"StarRailCard/**/*.py" = ["ALL"]
"ENCard/**/*.py" = ["ALL"]
"enka_card/**/*.py" = ["ALL"]

[tool.ruff.lint.flake8-type-checking]
quote-annotations = true


[tool.pyright]
exclude = [
    "EnkaCard/**/*.py",
    ".venv/**/*.py",
    ".venv/**/*.pyi",
    "StarRailCard/**/*.py",
    "ENCard/**/*.py",
    "enka_card/**/*.py"
]